Rename School to SchoolPage
