+++ src/Pages/SchoolPage.js
